Restructured the code in order to have each functionality in each module.
This wasn't made previously as I find it quicker to work with every 
functionality in the same file and, because i work alone, i don't need 
to worry about colleagues having trouble reading it.

Since im not planing on working on it anymore, and in order to make it 
more readable to other people, I'm starting the restructure here.